tracing_pillars_of_observability_tutorial/requirements.txt
flask
opentelemetry-exporter-jaeger
opentelemetry-sdk
requests
